packages/backend/requirements.txt
Flask==1.1.1
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.13
psycopg2==2.8.6
marshmallow==3.4.0
gunicorn==19.9.0
py3-validate-email==0.2.12
pycountry==20.7.3
